+++ libs/db/remotedb/proto/defs.pb.go
@@ -19,7 +19,7 @@ It has these top-level messages:
 */
 package protodb
 
-import proto "github.com/golang/protobuf/proto"
+import proto "github.com/gracenoah/golangprotobuf/proto"
 import fmt "fmt"
 import math "math"